@@ -1,101 +1,48 @@
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.5)
 project(vesc_ackermann)
 
-# Setting C++ standard to 11
-if (NOT "${CMAKE_CXX_STANDARD_COMPUTED_DEFAULT}")
-  message(STATUS "Changing CXX_STANDARD from C++98 to C++11")
-  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
-elseif ("${CMAKE_CXX_STANDARD_COMPUTED_DEFAULT}" STREQUAL "98")
-  message(STATUS "Changing CXX_STANDARD from C++98 to C++11")
-  set(CMAKE_CXX_STANDARD 11)
+# Setting C++ standard to 14
+if(NOT "${CMAKE_CXX_STANDARD_COMPUTED_DEFAULT}")
+  message(STATUS "Changing CXX_STANDARD from C++98 to C++14")
+  set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++14")
+elseif("${CMAKE_CXX_STANDARD_COMPUTED_DEFAULT}" STREQUAL "98")
+  message(STATUS "Changing CXX_STANDARD from C++98 to C++14")
+  set(CMAKE_CXX_STANDARD 14)
 endif()
 
-find_package(catkin REQUIRED COMPONENTS
-  ackermann_msgs
-  geometry_msgs
-  nav_msgs
-  nodelet
-  pluginlib
-  roscpp
-  roslint
-  std_msgs
-  tf
-  vesc_msgs
-)
-
-catkin_package(
-  INCLUDE_DIRS include
-  CATKIN_DEPENDS
-    ackermann_msgs
-    geometry_msgs
-    nav_msgs
-    nodelet
-    pluginlib
-    std_msgs
-    tf
-    vesc_msgs
-)
+find_package(ament_cmake_auto REQUIRED)
+ament_auto_find_build_dependencies()
 
 ###########
 ## Build ##
 ###########
 
-include_directories(
-  include
-  ${catkin_INCLUDE_DIRS}
+# nodes library
+ament_auto_add_library(${PROJECT_NAME} SHARED
+  src/ackermann_to_vesc.cpp
+  src/vesc_to_odom.cpp
 )
 
-# node executable
-add_executable(vesc_to_odom_node src/vesc_to_odom_node.cpp
-                                 src/vesc_to_odom.cpp)
-add_dependencies(vesc_to_odom_node ${catkin_EXPORTED_TARGETS})
-target_link_libraries(vesc_to_odom_node
-  ${catkin_LIBRARIES}
+# register nodes as components
+rclcpp_components_register_node(${PROJECT_NAME}
+  PLUGIN vesc_ackermann::AckermannToVesc
+  EXECUTABLE ackermann_to_vesc_node
 )
-
-add_executable(ackermann_to_vesc_node src/ackermann_to_vesc_node.cpp
-                                      src/ackermann_to_vesc.cpp)
-add_dependencies(ackermann_to_vesc_node ${catkin_EXPORTED_TARGETS})
-target_link_libraries(ackermann_to_vesc_node
-  ${catkin_LIBRARIES}
+rclcpp_components_register_node(${PROJECT_NAME}
+  PLUGIN vesc_ackermann::VescToOdom
+  EXECUTABLE vesc_to_odom_node
 )
-
-# nodelet library
-add_library(vesc_ackermann_nodelet src/ackermann_to_vesc_nodelet.cpp
-                                   src/ackermann_to_vesc.cpp
-                                   src/vesc_to_odom_nodelet.cpp
-                                   src/vesc_to_odom.cpp)
-add_dependencies(vesc_ackermann_nodelet ${catkin_EXPORTED_TARGETS})
-target_link_libraries(vesc_ackermann_nodelet
-  ${catkin_LIBRARIES}
-)
-
-set(ROSLINT_CPP_OPTS "--filter=-build/c++11")
-roslint_cpp()
-
-#############
-## Install ##
-#############
-
-install(TARGETS vesc_to_odom_node ackermann_to_vesc_node
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
-)
-
-install(TARGETS vesc_ackermann_nodelet
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-)
-
-install(DIRECTORY include/${PROJECT_NAME}/
-        DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION})
-install(FILES vesc_ackermann_nodelet.xml
-        DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
-install(DIRECTORY launch/
-        DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION}/launch)
 
 #############
 ## Testing ##
 #############
 
-if(CATKIN_ENABLE_TESTING)
-  roslint_add_test()
+if(BUILD_TESTING)
+  find_package(ament_lint_auto REQUIRED)
+  ament_lint_auto_find_test_dependencies()
 endif()
+
+ament_auto_package(
+  INSTALL_TO_SHARE
+    launch
+)
