@@ -1,188 +1,190 @@
 // Copyright 2020 F1TENTH Foundation
 //
-// Redistribution and use in source and binary forms, with or without modification, are permitted
-// provided that the following conditions are met:
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
 //
-// 1. Redistributions of source code must retain the above copyright notice, this list of conditions
-//    and the following disclaimer.
+//   * Redistributions of source code must retain the above copyright
+//     notice, this list of conditions and the following disclaimer.
 //
-// 2. Redistributions in binary form must reproduce the above copyright notice, this list
-//    of conditions and the following disclaimer in the documentation and/or other materials
-//    provided with the distribution.
+//   * Redistributions in binary form must reproduce the above copyright
+//     notice, this list of conditions and the following disclaimer in the
+//     documentation and/or other materials provided with the distribution.
 //
-// 3. Neither the name of the copyright holder nor the names of its contributors may be used
-//    to endorse or promote products derived from this software without specific prior
-//    written permission.
+//   * Neither the name of the {copyright_holder} nor the names of its
+//     contributors may be used to endorse or promote products derived from
+//     this software without specific prior written permission.
 //
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
-// IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
-// FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
-// WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
 
 // -*- mode:c++; fill-column: 100; -*-
 
-#include "vesc_ackermann/vesc_to_odom.h"
+#include "vesc_ackermann/vesc_to_odom.hpp"
+
+#include <geometry_msgs/msg/transform_stamped.hpp>
+#include <vesc_msgs/msg/vesc_state_stamped.hpp>
 
 #include <cmath>
 #include <string>
 
-#include <nav_msgs/Odometry.h>
-#include <geometry_msgs/TransformStamped.h>
-
 namespace vesc_ackermann
 {
 
-template <typename T>
-inline bool getRequiredParam(const ros::NodeHandle& nh, const std::string& name, T* value);
+using geometry_msgs::msg::TransformStamped;
+using nav_msgs::msg::Odometry;
+using std::placeholders::_1;
+using std_msgs::msg::Float64;
+using vesc_msgs::msg::VescStateStamped;
 
-VescToOdom::VescToOdom(ros::NodeHandle nh, ros::NodeHandle private_nh) :
-  odom_frame_("odom"), base_frame_("base_link"),
-  use_servo_cmd_(true), publish_tf_(false), x_(0.0), y_(0.0), yaw_(0.0)
+VescToOdom::VescToOdom(const rclcpp::NodeOptions & options)
+: Node("vesc_to_odom_node", options),
+  odom_frame_("odom"),
+  base_frame_("base_link"),
+  use_servo_cmd_(true),
+  publish_tf_(false),
+  x_(0.0),
+  y_(0.0),
+  yaw_(0.0)
 {
   // get ROS parameters
-  private_nh.param("odom_frame", odom_frame_, odom_frame_);
-  private_nh.param("base_frame", base_frame_, base_frame_);
-  private_nh.param("use_servo_cmd_to_calc_angular_velocity", use_servo_cmd_, use_servo_cmd_);
-  if (!getRequiredParam(nh, "speed_to_erpm_gain", &speed_to_erpm_gain_))
-    return;
-  if (!getRequiredParam(nh, "speed_to_erpm_offset", &speed_to_erpm_offset_))
-    return;
-  if (use_servo_cmd_)
-  {
-    if (!getRequiredParam(nh, "steering_angle_to_servo_gain", &steering_to_servo_gain_))
-      return;
-    if (!getRequiredParam(nh, "steering_angle_to_servo_offset", &steering_to_servo_offset_))
-      return;
-    if (!getRequiredParam(nh, "wheelbase", &wheelbase_))
-      return;
+  odom_frame_ = declare_parameter("odom_frame", odom_frame_);
+  base_frame_ = declare_parameter("base_frame", base_frame_);
+  use_servo_cmd_ = declare_parameter("use_servo_cmd_to_calc_angular_velocity", use_servo_cmd_);
+
+  speed_to_erpm_gain_ = declare_parameter("speed_to_erpm_gain").get<double>();
+  speed_to_erpm_offset_ = declare_parameter("speed_to_erpm_offset").get<double>();
+
+  if (use_servo_cmd_) {
+    steering_to_servo_gain_ = declare_parameter("steering_angle_to_servo_gain").get<double>();
+    steering_to_servo_offset_ = declare_parameter("steering_angle_to_servo_offset").get<double>();
+    wheelbase_ = declare_parameter("wheelbase").get<double>();
   }
-  private_nh.param("publish_tf", publish_tf_, publish_tf_);
+
+  publish_tf_ = declare_parameter("publish_tf", publish_tf_);
 
   // create odom publisher
-  odom_pub_ = nh.advertise<nav_msgs::Odometry>("odom", 10);
+  odom_pub_ = create_publisher<Odometry>("odom", 10);
 
   // create tf broadcaster
-  if (publish_tf_)
-  {
-    tf_pub_.reset(new tf::TransformBroadcaster);
+  if (publish_tf_) {
+    tf_pub_.reset(new tf2_ros::TransformBroadcaster(this));
   }
 
   // subscribe to vesc state and. optionally, servo command
-  vesc_state_sub_ = nh.subscribe("sensors/core", 10, &VescToOdom::vescStateCallback, this);
-  if (use_servo_cmd_)
-  {
-    servo_sub_ = nh.subscribe("sensors/servo_position_command", 10,
-                              &VescToOdom::servoCmdCallback, this);
+  vesc_state_sub_ = create_subscription<VescStateStamped>(
+    "sensors/core", 10, std::bind(&VescToOdom::vescStateCallback, this, _1));
+
+  if (use_servo_cmd_) {
+    servo_sub_ = create_subscription<Float64>(
+      "sensors/servo_position_command", 10, std::bind(&VescToOdom::servoCmdCallback, this, _1));
   }
 }
 
-void VescToOdom::vescStateCallback(const vesc_msgs::VescStateStamped::ConstPtr& state)
+void VescToOdom::vescStateCallback(const VescStateStamped::SharedPtr state)
 {
   // check that we have a last servo command if we are depending on it for angular velocity
-  if (use_servo_cmd_ && !last_servo_cmd_)
+  if (use_servo_cmd_ && !last_servo_cmd_) {
     return;
+  }
 
   // convert to engineering units
   double current_speed = (-state->state.speed - speed_to_erpm_offset_) / speed_to_erpm_gain_;
-  if (std::fabs(current_speed) < 0.05)
-  {
+  if (std::fabs(current_speed) < 0.05) {
     current_speed = 0.0;
   }
   double current_steering_angle(0.0), current_angular_velocity(0.0);
-  if (use_servo_cmd_)
-  {
+  if (use_servo_cmd_) {
     current_steering_angle =
       (last_servo_cmd_->data - steering_to_servo_offset_) / steering_to_servo_gain_;
     current_angular_velocity = current_speed * tan(current_steering_angle) / wheelbase_;
   }
 
   // use current state as last state if this is our first time here
-  if (!last_state_)
+  if (!last_state_) {
     last_state_ = state;
+  }
 
   // calc elapsed time
-  ros::Duration dt = state->header.stamp - last_state_->header.stamp;
+  auto dt = rclcpp::Time(state->header.stamp) - rclcpp::Time(last_state_->header.stamp);
 
   /** @todo could probably do better propigating odometry, e.g. trapezoidal integration */
 
   // propigate odometry
   double x_dot = current_speed * cos(yaw_);
   double y_dot = current_speed * sin(yaw_);
-  x_ += x_dot * dt.toSec();
-  y_ += y_dot * dt.toSec();
-  if (use_servo_cmd_)
-    yaw_ += current_angular_velocity * dt.toSec();
+  x_ += x_dot * dt.seconds();
+  y_ += y_dot * dt.seconds();
+  if (use_servo_cmd_) {
+    yaw_ += current_angular_velocity * dt.seconds();
+  }
 
   // save state for next time
   last_state_ = state;
 
   // publish odometry message
-  nav_msgs::Odometry::Ptr odom(new nav_msgs::Odometry);
-  odom->header.frame_id = odom_frame_;
-  odom->header.stamp = state->header.stamp;
-  odom->child_frame_id = base_frame_;
+  Odometry odom;
+  odom.header.frame_id = odom_frame_;
+  odom.header.stamp = state->header.stamp;
+  odom.child_frame_id = base_frame_;
 
   // Position
-  odom->pose.pose.position.x = x_;
-  odom->pose.pose.position.y = y_;
-  odom->pose.pose.orientation.x = 0.0;
-  odom->pose.pose.orientation.y = 0.0;
-  odom->pose.pose.orientation.z = sin(yaw_ / 2.0);
-  odom->pose.pose.orientation.w = cos(yaw_ / 2.0);
+  odom.pose.pose.position.x = x_;
+  odom.pose.pose.position.y = y_;
+  odom.pose.pose.orientation.x = 0.0;
+  odom.pose.pose.orientation.y = 0.0;
+  odom.pose.pose.orientation.z = sin(yaw_ / 2.0);
+  odom.pose.pose.orientation.w = cos(yaw_ / 2.0);
 
   // Position uncertainty
   /** @todo Think about position uncertainty, perhaps get from parameters? */
-  odom->pose.covariance[0]  = 0.2;  ///< x
-  odom->pose.covariance[7]  = 0.2;  ///< y
-  odom->pose.covariance[35] = 0.4;  ///< yaw
+  odom.pose.covariance[0] = 0.2;   ///< x
+  odom.pose.covariance[7] = 0.2;   ///< y
+  odom.pose.covariance[35] = 0.4;  ///< yaw
 
   // Velocity ("in the coordinate frame given by the child_frame_id")
-  odom->twist.twist.linear.x = current_speed;
-  odom->twist.twist.linear.y = 0.0;
-  odom->twist.twist.angular.z = current_angular_velocity;
+  odom.twist.twist.linear.x = current_speed;
+  odom.twist.twist.linear.y = 0.0;
+  odom.twist.twist.angular.z = current_angular_velocity;
 
   // Velocity uncertainty
   /** @todo Think about velocity uncertainty */
 
-  if (publish_tf_)
-  {
-    geometry_msgs::TransformStamped tf;
+  if (publish_tf_) {
+    TransformStamped tf;
     tf.header.frame_id = odom_frame_;
     tf.child_frame_id = base_frame_;
-    tf.header.stamp = ros::Time::now();
+    tf.header.stamp = now();
     tf.transform.translation.x = x_;
     tf.transform.translation.y = y_;
     tf.transform.translation.z = 0.0;
-    tf.transform.rotation = odom->pose.pose.orientation;
-    if (ros::ok())
-    {
+    tf.transform.rotation = odom.pose.pose.orientation;
+
+    if (rclcpp::ok()) {
       tf_pub_->sendTransform(tf);
     }
   }
 
-  if (ros::ok())
-  {
-    odom_pub_.publish(odom);
+  if (rclcpp::ok()) {
+    odom_pub_->publish(odom);
   }
 }
 
-void VescToOdom::servoCmdCallback(const std_msgs::Float64::ConstPtr& servo)
+void VescToOdom::servoCmdCallback(const Float64::SharedPtr servo)
 {
   last_servo_cmd_ = servo;
 }
 
-template <typename T>
-inline bool getRequiredParam(const ros::NodeHandle& nh, const std::string& name, T* value)
-{
-  if (nh.getParam(name, *value))
-    return true;
-
-  ROS_FATAL("VescToOdom: Parameter %s is required.", name.c_str());
-  return false;
-}
-
 }  // namespace vesc_ackermann
+
+#include "rclcpp_components/register_node_macro.hpp"  // NOLINT
+
+RCLCPP_COMPONENTS_REGISTER_NODE(vesc_ackermann::VescToOdom)
