@@ -1,38 +1,36 @@
+// Copyright (c) 2020, Daniel Bahr
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//   * Redistributions of source code must retain the above copyright
+//     notice, this list of conditions and the following disclaimer.
+//
+//   * Redistributions in binary form must reproduce the above copyright
+//     notice, this list of conditions and the following disclaimer in the
+//     documentation and/or other materials provided with the distribution.
+//
+//   * Neither the name of the {copyright_holder} nor the names of its
+//     contributors may be used to endorse or promote products derived from
+//     this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
 /**
-    @file CRC.h
+    @file CRC.hpp
     @author Daniel Bahr
     @version 1.0.1.0
     @copyright
-    @parblock
-        CRC++
-        Copyright (c) 2020, Daniel Bahr
-        All rights reserved.
-
-        Redistribution and use in source and binary forms, with or without
-        modification, are permitted provided that the following conditions are met:
-
-        * Redistributions of source code must retain the above copyright notice, this
-          list of conditions and the following disclaimer.
-
-        * Redistributions in binary form must reproduce the above copyright notice,
-          this list of conditions and the following disclaimer in the documentation
-          and/or other materials provided with the distribution.
-
-        * Neither the name of CRC++ nor the names of its
-          contributors may be used to endorse or promote products derived from
-          this software without specific prior written permission.
-
-        THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-        AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-        IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-        DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-        FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-        DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-        SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-        CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-        OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-        OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-    @endparblock
 */
 
 /*
@@ -52,78 +50,78 @@
                                                           multiplication in the bit-by-bit calculation instead of a small conditional. The branchless implementation
                                                           may be faster on processor architectures which support single-instruction integer multiplication.
         #define CRCPP_USE_CPP11                         - Define to enables C++11 features (move semantics, constexpr, static_assert, etc.).
-        #define CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS  - Define to include definitions for little-used CRCs. 
+        #define CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS  - Define to include definitions for little-used CRCs.
 */
 
-#ifndef VESC_DRIVER_CRC_H_
-#define VESC_DRIVER_CRC_H_
+#ifndef VESC_DRIVER__CRC_HPP_
+#define VESC_DRIVER__CRC_HPP_
 
-#include <climits>  // Includes CHAR_BIT
+#include <climits>   // Includes CHAR_BIT
 #ifdef CRCPP_USE_CPP11
-#include <cstddef>  // Includes ::std::size_t
-#include <cstdint>  // Includes ::std::uint8_t, ::std::uint16_t, ::std::uint32_t, ::std::uint64_t
+#include <cstddef>   // Includes ::std::size_t
+#include <cstdint>   // Includes ::std::uint8_t, ::std::uint16_t, ::std::uint32_t, ::std::uint64_t
 #else
-#include <stddef.h> // Includes size_t
-#include <stdint.h> // Includes uint8_t, uint16_t, uint32_t, uint64_t
+#include <stddef.h>  // Includes size_t
+#include <stdint.h>  // Includes uint8_t, uint16_t, uint32_t, uint64_t
 #endif
-#include <limits>   // Includes ::std::numeric_limits
-#include <utility>  // Includes ::std::move
+#include <limits>    // Includes ::std::numeric_limits
+#include <utility>   // Includes ::std::move
 
 #ifndef crcpp_uint8
 #   ifdef CRCPP_USE_CPP11
-        /// @brief Unsigned 8-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 8-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint8 ::std::uint8_t
 #   else
-        /// @brief Unsigned 8-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 8-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint8 uint8_t
 #   endif
 #endif
 
 #ifndef crcpp_uint16
 #   ifdef CRCPP_USE_CPP11
-        /// @brief Unsigned 16-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 16-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint16 ::std::uint16_t
 #   else
-        /// @brief Unsigned 16-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 16-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint16 uint16_t
 #   endif
 #endif
 
 #ifndef crcpp_uint32
 #   ifdef CRCPP_USE_CPP11
-        /// @brief Unsigned 32-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 32-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint32 ::std::uint32_t
 #   else
-        /// @brief Unsigned 32-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 32-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint32 uint32_t
 #   endif
 #endif
 
 #ifndef crcpp_uint64
 #   ifdef CRCPP_USE_CPP11
-        /// @brief Unsigned 64-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 64-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint64 ::std::uint64_t
 #   else
-        /// @brief Unsigned 64-bit integer definition, used primarily for parameter definitions.
+/// @brief Unsigned 64-bit integer definition, used primarily for parameter definitions.
 #       define crcpp_uint64 uint64_t
 #   endif
 #endif
 
 #ifndef crcpp_size
 #   ifdef CRCPP_USE_CPP11
-        /// @brief Unsigned size definition, used for specifying data sizes.
+/// @brief Unsigned size definition, used for specifying data sizes.
 #       define crcpp_size ::std::size_t
 #   else
-        /// @brief Unsigned size definition, used for specifying data sizes.
+/// @brief Unsigned size definition, used for specifying data sizes.
 #       define crcpp_size size_t
 #   endif
 #endif
 
 #ifdef CRCPP_USE_CPP11
-    /// @brief Compile-time expression definition.
+/// @brief Compile-time expression definition.
 #   define crcpp_constexpr constexpr
 #else
-    /// @brief Compile-time expression definition.
+/// @brief Compile-time expression definition.
 #   define crcpp_constexpr const
 #endif
 
@@ -142,163 +140,176 @@ namespace CRCPP
 class CRC
 {
 public:
-    // Forward declaration
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    struct Table;
+  // Forward declaration
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  struct Table;
 
-    /**
-        @brief CRC parameters.
-    */
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    struct Parameters
-    {
-        CRCType polynomial;   ///< CRC polynomial
-        CRCType initialValue; ///< Initial CRC value
-        CRCType finalXOR;     ///< Value to XOR with the final CRC
-        bool reflectInput;    ///< true to reflect all input bytes
-        bool reflectOutput;   ///< true to reflect the output CRC (reflection occurs before the final XOR)
+  /**
+      @brief CRC parameters.
+  */
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  struct Parameters
+  {
+    CRCType polynomial;       ///< CRC polynomial
+    CRCType initialValue;     ///< Initial CRC value
+    CRCType finalXOR;         ///< Value to XOR with the final CRC
+    bool reflectInput;        ///< true to reflect all input bytes
+    bool reflectOutput;       ///< true to reflect the output CRC (reflection occurs before the final XOR)  // NOLINT
 
-        Table<CRCType, CRCWidth> MakeTable() const;
-    };
+    Table<CRCType, CRCWidth> MakeTable() const;
+  };
 
-    /**
-        @brief CRC lookup table. After construction, the CRC parameters are fixed.
-        @note A CRC table can be used for multiple CRC calculations.
-    */
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    struct Table
-    {
-        // Constructors are intentionally NOT marked explicit.
-        Table(const Parameters<CRCType, CRCWidth> & parameters);
-
-#ifdef CRCPP_USE_CPP11
-        Table(Parameters<CRCType, CRCWidth> && parameters);
-#endif
-
-        const Parameters<CRCType, CRCWidth> & GetParameters() const;
-
-        const CRCType * GetTable() const;
-
-        CRCType operator[](unsigned char index) const;
-
-    private:
-        void InitTable();
-
-        Parameters<CRCType, CRCWidth> parameters; ///< CRC parameters used to construct the table
-        CRCType table[1 << CHAR_BIT];             ///< CRC lookup table
-    };
-
-    // The number of bits in CRCType must be at least as large as CRCWidth.
-    // CRCType must be an unsigned integer type or a custom type with operator overloads.
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType Calculate(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters);
-
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType Calculate(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters, CRCType crc);
-
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType Calculate(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable);
-
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType Calculate(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable, CRCType crc);
-
-    // Common CRCs up to 64 bits.
-    // Note: Check values are the computed CRCs when given an ASCII input of "123456789" (without null terminator)
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters< crcpp_uint8,  4> & CRC_4_ITU();
-    static const Parameters< crcpp_uint8,  5> & CRC_5_EPC();
-    static const Parameters< crcpp_uint8,  5> & CRC_5_ITU();
-    static const Parameters< crcpp_uint8,  5> & CRC_5_USB();
-    static const Parameters< crcpp_uint8,  6> & CRC_6_CDMA2000A();
-    static const Parameters< crcpp_uint8,  6> & CRC_6_CDMA2000B();
-    static const Parameters< crcpp_uint8,  6> & CRC_6_ITU();
-    static const Parameters< crcpp_uint8,  7> & CRC_7();
-#endif
-    static const Parameters< crcpp_uint8,  8> & CRC_8();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters< crcpp_uint8,  8> & CRC_8_EBU();
-    static const Parameters< crcpp_uint8,  8> & CRC_8_MAXIM();
-    static const Parameters< crcpp_uint8,  8> & CRC_8_WCDMA();
-    static const Parameters<crcpp_uint16, 10> & CRC_10();
-    static const Parameters<crcpp_uint16, 10> & CRC_10_CDMA2000();
-    static const Parameters<crcpp_uint16, 11> & CRC_11();
-    static const Parameters<crcpp_uint16, 12> & CRC_12_CDMA2000();
-    static const Parameters<crcpp_uint16, 12> & CRC_12_DECT();
-    static const Parameters<crcpp_uint16, 12> & CRC_12_UMTS();
-    static const Parameters<crcpp_uint16, 13> & CRC_13_BBC();
-    static const Parameters<crcpp_uint16, 15> & CRC_15();
-    static const Parameters<crcpp_uint16, 15> & CRC_15_MPT1327();
-#endif
-    static const Parameters<crcpp_uint16, 16> & CRC_16_ARC();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_BUYPASS();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_CCITTFALSE();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters<crcpp_uint16, 16> & CRC_16_CDMA2000();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_CMS();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_DECTR();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_DECTX();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_DNP();
-#endif
-    static const Parameters<crcpp_uint16, 16> & CRC_16_GENIBUS();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_KERMIT();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters<crcpp_uint16, 16> & CRC_16_MAXIM();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_MODBUS();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_T10DIF();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_USB();
-#endif
-    static const Parameters<crcpp_uint16, 16> & CRC_16_X25();
-    static const Parameters<crcpp_uint16, 16> & CRC_16_XMODEM();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters<crcpp_uint32, 17> & CRC_17_CAN();
-    static const Parameters<crcpp_uint32, 21> & CRC_21_CAN();
-    static const Parameters<crcpp_uint32, 24> & CRC_24();
-    static const Parameters<crcpp_uint32, 24> & CRC_24_FLEXRAYA();
-    static const Parameters<crcpp_uint32, 24> & CRC_24_FLEXRAYB();
-    static const Parameters<crcpp_uint32, 30> & CRC_30();
-#endif
-    static const Parameters<crcpp_uint32, 32> & CRC_32();
-    static const Parameters<crcpp_uint32, 32> & CRC_32_BZIP2();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters<crcpp_uint32, 32> & CRC_32_C();
-#endif
-    static const Parameters<crcpp_uint32, 32> & CRC_32_MPEG2();
-    static const Parameters<crcpp_uint32, 32> & CRC_32_POSIX();
-#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
-    static const Parameters<crcpp_uint32, 32> & CRC_32_Q();
-    static const Parameters<crcpp_uint64, 40> & CRC_40_GSM();
-    static const Parameters<crcpp_uint64, 64> & CRC_64();
-#endif
+  /**
+      @brief CRC lookup table. After construction, the CRC parameters are fixed.
+      @note A CRC table can be used for multiple CRC calculations.
+  */
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  struct Table
+  {
+    // Constructors are intentionally NOT marked explicit.
+    explicit Table(const Parameters<CRCType, CRCWidth> & parameters);
 
 #ifdef CRCPP_USE_CPP11
-    CRC() = delete;
-    CRC(const CRC & other) = delete;
-    CRC & operator=(const CRC & other) = delete;
-    CRC(CRC && other) = delete;
-    CRC & operator=(CRC && other) = delete;
+    explicit Table(Parameters<CRCType, CRCWidth> && parameters);
+#endif
+
+    const Parameters<CRCType, CRCWidth> & GetParameters() const;
+
+    const CRCType * GetTable() const;
+
+    CRCType operator[](unsigned char index) const;
+
+private:
+    void InitTable();
+
+    Parameters<CRCType, CRCWidth> parameters;     ///< CRC parameters used to construct the table
+    CRCType table[1 << CHAR_BIT];                 ///< CRC lookup table
+  };
+
+  // The number of bits in CRCType must be at least as large as CRCWidth.
+  // CRCType must be an unsigned integer type or a custom type with operator overloads.
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType Calculate(
+    const void * data, crcpp_size size, const Parameters<CRCType,
+    CRCWidth> & parameters);
+
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType Calculate(
+    const void * data, crcpp_size size, const Parameters<CRCType,
+    CRCWidth> & parameters, CRCType crc);
+
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType Calculate(
+    const void * data, crcpp_size size, const Table<CRCType,
+    CRCWidth> & lookupTable);
+
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType Calculate(
+    const void * data, crcpp_size size, const Table<CRCType,
+    CRCWidth> & lookupTable, CRCType crc);
+
+  // Common CRCs up to 64 bits.
+  // Note: Check values are the computed CRCs when given an ASCII input of "123456789"
+  // (without null terminator)
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint8, 4> & CRC_4_ITU();
+  static const Parameters<crcpp_uint8, 5> & CRC_5_EPC();
+  static const Parameters<crcpp_uint8, 5> & CRC_5_ITU();
+  static const Parameters<crcpp_uint8, 5> & CRC_5_USB();
+  static const Parameters<crcpp_uint8, 6> & CRC_6_CDMA2000A();
+  static const Parameters<crcpp_uint8, 6> & CRC_6_CDMA2000B();
+  static const Parameters<crcpp_uint8, 6> & CRC_6_ITU();
+  static const Parameters<crcpp_uint8, 7> & CRC_7();
+#endif
+  static const Parameters<crcpp_uint8, 8> & CRC_8();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint8, 8> & CRC_8_EBU();
+  static const Parameters<crcpp_uint8, 8> & CRC_8_MAXIM();
+  static const Parameters<crcpp_uint8, 8> & CRC_8_WCDMA();
+  static const Parameters<crcpp_uint16, 10> & CRC_10();
+  static const Parameters<crcpp_uint16, 10> & CRC_10_CDMA2000();
+  static const Parameters<crcpp_uint16, 11> & CRC_11();
+  static const Parameters<crcpp_uint16, 12> & CRC_12_CDMA2000();
+  static const Parameters<crcpp_uint16, 12> & CRC_12_DECT();
+  static const Parameters<crcpp_uint16, 12> & CRC_12_UMTS();
+  static const Parameters<crcpp_uint16, 13> & CRC_13_BBC();
+  static const Parameters<crcpp_uint16, 15> & CRC_15();
+  static const Parameters<crcpp_uint16, 15> & CRC_15_MPT1327();
+#endif
+  static const Parameters<crcpp_uint16, 16> & CRC_16_ARC();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_BUYPASS();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_CCITTFALSE();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint16, 16> & CRC_16_CDMA2000();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_CMS();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_DECTR();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_DECTX();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_DNP();
+#endif
+  static const Parameters<crcpp_uint16, 16> & CRC_16_GENIBUS();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_KERMIT();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint16, 16> & CRC_16_MAXIM();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_MODBUS();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_T10DIF();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_USB();
+#endif
+  static const Parameters<crcpp_uint16, 16> & CRC_16_X25();
+  static const Parameters<crcpp_uint16, 16> & CRC_16_XMODEM();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint32, 17> & CRC_17_CAN();
+  static const Parameters<crcpp_uint32, 21> & CRC_21_CAN();
+  static const Parameters<crcpp_uint32, 24> & CRC_24();
+  static const Parameters<crcpp_uint32, 24> & CRC_24_FLEXRAYA();
+  static const Parameters<crcpp_uint32, 24> & CRC_24_FLEXRAYB();
+  static const Parameters<crcpp_uint32, 30> & CRC_30();
+#endif
+  static const Parameters<crcpp_uint32, 32> & CRC_32();
+  static const Parameters<crcpp_uint32, 32> & CRC_32_BZIP2();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint32, 32> & CRC_32_C();
+#endif
+  static const Parameters<crcpp_uint32, 32> & CRC_32_MPEG2();
+  static const Parameters<crcpp_uint32, 32> & CRC_32_POSIX();
+#ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+  static const Parameters<crcpp_uint32, 32> & CRC_32_Q();
+  static const Parameters<crcpp_uint64, 40> & CRC_40_GSM();
+  static const Parameters<crcpp_uint64, 64> & CRC_64();
+#endif
+
+#ifdef CRCPP_USE_CPP11
+  CRC() = delete;
+  CRC(const CRC & other) = delete;
+  CRC & operator=(const CRC & other) = delete;
+  CRC(CRC && other) = delete;
+  CRC & operator=(CRC && other) = delete;
 #endif
 
 private:
 #ifndef CRCPP_USE_CPP11
-    CRC();
-    CRC(const CRC & other);
-    CRC & operator=(const CRC & other);
+  CRC();
+  CRC(const CRC & other);
+  CRC & operator=(const CRC & other);
 #endif
 
-    template <typename IntegerType>
-    static IntegerType Reflect(IntegerType value, crcpp_uint16 numBits);
+  template<typename IntegerType>
+  static IntegerType Reflect(IntegerType value, crcpp_uint16 numBits);
 
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType Finalize(CRCType remainder, CRCType finalXOR, bool reflectOutput);
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType Finalize(CRCType remainder, CRCType finalXOR, bool reflectOutput);
 
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType UndoFinalize(CRCType remainder, CRCType finalXOR, bool reflectOutput);
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType UndoFinalize(CRCType remainder, CRCType finalXOR, bool reflectOutput);
 
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType CalculateRemainder(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters, CRCType remainder);
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType CalculateRemainder(
+    const void * data, crcpp_size size, const Parameters<CRCType,
+    CRCWidth> & parameters, CRCType remainder);
 
-    template <typename CRCType, crcpp_uint16 CRCWidth>
-    static CRCType CalculateRemainder(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable, CRCType remainder);
+  template<typename CRCType, crcpp_uint16 CRCWidth>
+  static CRCType CalculateRemainder(
+    const void * data, crcpp_size size, const Table<CRCType,
+    CRCWidth> & lookupTable, CRCType remainder);
 };
 
 /**
@@ -309,11 +320,11 @@ private:
     @tparam CRCWidth Number of bits in the CRC
     @return CRC lookup table
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline CRC::Table<CRCType, CRCWidth> CRC::Parameters<CRCType, CRCWidth>::MakeTable() const
 {
-    // This should take advantage of RVO and optimize out the copy.
-    return CRC::Table<CRCType, CRCWidth>(*this);
+  // This should take advantage of RVO and optimize out the copy.
+  return CRC::Table<CRCType, CRCWidth>(*this);
 }
 
 /**
@@ -322,11 +333,11 @@ inline CRC::Table<CRCType, CRCWidth> CRC::Parameters<CRCType, CRCWidth>::MakeTab
     @tparam CRCType Integer type for storing the CRC result
     @tparam CRCWidth Number of bits in the CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRC::Table<CRCType, CRCWidth>::Table(const Parameters<CRCType, CRCWidth> & params) :
-    parameters(params)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRC::Table<CRCType, CRCWidth>::Table(const Parameters<CRCType, CRCWidth> & params)
+: parameters(params)
 {
-    InitTable();
+  InitTable();
 }
 
 #ifdef CRCPP_USE_CPP11
@@ -336,11 +347,11 @@ inline CRC::Table<CRCType, CRCWidth>::Table(const Parameters<CRCType, CRCWidth> 
     @tparam CRCType Integer type for storing the CRC result
     @tparam CRCWidth Number of bits in the CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRC::Table<CRCType, CRCWidth>::Table(Parameters<CRCType, CRCWidth> && params) :
-    parameters(::std::move(params))
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRC::Table<CRCType, CRCWidth>::Table(Parameters<CRCType, CRCWidth> && params)
+: parameters(::std::move(params))
 {
-    InitTable();
+  InitTable();
 }
 #endif
 
@@ -350,10 +361,11 @@ inline CRC::Table<CRCType, CRCWidth>::Table(Parameters<CRCType, CRCWidth> && par
     @tparam CRCWidth Number of bits in the CRC
     @return CRC parameters
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline const CRC::Parameters<CRCType, CRCWidth> & CRC::Table<CRCType, CRCWidth>::GetParameters() const
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline const CRC::Parameters<CRCType, CRCWidth> & CRC::Table<CRCType,
+  CRCWidth>::GetParameters() const
 {
-    return parameters;
+  return parameters;
 }
 
 /**
@@ -362,10 +374,10 @@ inline const CRC::Parameters<CRCType, CRCWidth> & CRC::Table<CRCType, CRCWidth>:
     @tparam CRCWidth Number of bits in the CRC
     @return CRC table
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline const CRCType * CRC::Table<CRCType, CRCWidth>::GetTable() const
 {
-    return table;
+  return table;
 }
 
 /**
@@ -375,10 +387,10 @@ inline const CRCType * CRC::Table<CRCType, CRCWidth>::GetTable() const
     @tparam CRCWidth Number of bits in the CRC
     @return CRC table entry
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline CRCType CRC::Table<CRCType, CRCWidth>::operator[](unsigned char index) const
 {
-    return table[index];
+  return table[index];
 }
 
 /**
@@ -386,39 +398,39 @@ inline CRCType CRC::Table<CRCType, CRCWidth>::operator[](unsigned char index) co
     @tparam CRCType Integer type for storing the CRC result
     @tparam CRCWidth Number of bits in the CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline void CRC::Table<CRCType, CRCWidth>::InitTable()
 {
-    // For masking off the bits for the CRC (in the event that the number of bits in CRCType is larger than CRCWidth)
-    static crcpp_constexpr CRCType BIT_MASK((CRCType(1) << (CRCWidth - CRCType(1))) |
-                                           ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1)));
+  // For masking off the bits for the CRC
+  // (in the event that the number of bits in CRCType is larger than CRCWidth)
+  static crcpp_constexpr CRCType BIT_MASK((CRCType(1) << (CRCWidth - CRCType(1))) |
+    ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1)));
 
-    // The conditional expression is used to avoid a -Wshift-count-overflow warning.
-    static crcpp_constexpr CRCType SHIFT((CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
+  // The conditional expression is used to avoid a -Wshift-count-overflow warning.
+  static crcpp_constexpr CRCType SHIFT(
+    (CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
 
-    CRCType crc;
-    unsigned char byte = 0;
+  CRCType crc;
+  unsigned char byte = 0;
 
-    // Loop over each dividend (each possible number storable in an unsigned char)
-    do
-    {
-        crc = CRC::CalculateRemainder<CRCType, CRCWidth>(&byte, sizeof(byte), parameters, CRCType(0));
+  // Loop over each dividend (each possible number storable in an unsigned char)
+  do {
+    crc = CRC::CalculateRemainder<CRCType, CRCWidth>(&byte, sizeof(byte), parameters, CRCType(0));
 
-        // This mask might not be necessary; all unit tests pass with this line commented out,
-        // but that might just be a coincidence based on the CRC parameters used for testing.
-        // In any case, this is harmless to leave in and only adds a single machine instruction per loop iteration.
-        crc &= BIT_MASK;
+    // This mask might not be necessary; all unit tests pass with this line commented out,
+    // but that might just be a coincidence based on the CRC parameters used for testing.
+    // In any case, this is harmless to leave in and only adds a single machine instruction
+    // per loop iteration.
+    crc &= BIT_MASK;
 
-        if (!parameters.reflectInput && CRCWidth < CHAR_BIT)
-        {
-            // Undo the special operation at the end of the CalculateRemainder()
-            // function for non-reflected CRCs < CHAR_BIT.
-            crc = static_cast<CRCType>(crc << SHIFT);
-        }
-
-        table[byte] = crc;
+    if (!parameters.reflectInput && CRCWidth < CHAR_BIT) {
+      // Undo the special operation at the end of the CalculateRemainder()
+      // function for non-reflected CRCs < CHAR_BIT.
+      crc = static_cast<CRCType>(crc << SHIFT);
     }
-    while (++byte);
+
+    table[byte] = crc;
+  } while (++byte);
 }
 
 /**
@@ -430,14 +442,18 @@ inline void CRC::Table<CRCType, CRCWidth>::InitTable()
     @tparam CRCWidth Number of bits in the CRC
     @return CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::Calculate(
+  const void * data, crcpp_size size, const Parameters<CRCType,
+  CRCWidth> & parameters)
 {
-    CRCType remainder = CalculateRemainder(data, size, parameters, parameters.initialValue);
+  CRCType remainder = CalculateRemainder(data, size, parameters, parameters.initialValue);
 
-    // No need to mask the remainder here; the mask will be applied in the Finalize() function.
+  // No need to mask the remainder here; the mask will be applied in the Finalize() function.
 
-    return Finalize<CRCType, CRCWidth>(remainder, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  return Finalize<CRCType, CRCWidth>(
+    remainder, parameters.finalXOR,
+    parameters.reflectInput != parameters.reflectOutput);
 }
 /**
     @brief Appends additional data to a previous CRC calculation.
@@ -450,16 +466,23 @@ inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Paramete
     @tparam CRCWidth Number of bits in the CRC
     @return CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters, CRCType crc)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::Calculate(
+  const void * data, crcpp_size size, const Parameters<CRCType,
+  CRCWidth> & parameters, CRCType crc)
 {
-    CRCType remainder = UndoFinalize<CRCType, CRCWidth>(crc, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  CRCType remainder = UndoFinalize<CRCType, CRCWidth>(
+    crc, parameters.finalXOR,
+    parameters.reflectInput !=
+    parameters.reflectOutput);
 
-    remainder = CalculateRemainder(data, size, parameters, remainder);
+  remainder = CalculateRemainder(data, size, parameters, remainder);
 
-    // No need to mask the remainder here; the mask will be applied in the Finalize() function.
+  // No need to mask the remainder here; the mask will be applied in the Finalize() function.
 
-    return Finalize<CRCType, CRCWidth>(remainder, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  return Finalize<CRCType, CRCWidth>(
+    remainder, parameters.finalXOR,
+    parameters.reflectInput != parameters.reflectOutput);
 }
 
 /**
@@ -471,16 +494,20 @@ inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Paramete
     @tparam CRCWidth Number of bits in the CRC
     @return CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::Calculate(
+  const void * data, crcpp_size size, const Table<CRCType,
+  CRCWidth> & lookupTable)
 {
-    const Parameters<CRCType, CRCWidth> & parameters = lookupTable.GetParameters();
+  const Parameters<CRCType, CRCWidth> & parameters = lookupTable.GetParameters();
 
-    CRCType remainder = CalculateRemainder(data, size, lookupTable, parameters.initialValue);
+  CRCType remainder = CalculateRemainder(data, size, lookupTable, parameters.initialValue);
 
-    // No need to mask the remainder here; the mask will be applied in the Finalize() function.
+  // No need to mask the remainder here; the mask will be applied in the Finalize() function.
 
-    return Finalize<CRCType, CRCWidth>(remainder, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  return Finalize<CRCType, CRCWidth>(
+    remainder, parameters.finalXOR,
+    parameters.reflectInput != parameters.reflectOutput);
 }
 
 /**
@@ -494,18 +521,25 @@ inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Table<CR
     @tparam CRCWidth Number of bits in the CRC
     @return CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable, CRCType crc)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::Calculate(
+  const void * data, crcpp_size size, const Table<CRCType,
+  CRCWidth> & lookupTable, CRCType crc)
 {
-    const Parameters<CRCType, CRCWidth> & parameters = lookupTable.GetParameters();
+  const Parameters<CRCType, CRCWidth> & parameters = lookupTable.GetParameters();
 
-    CRCType remainder = UndoFinalize<CRCType, CRCWidth>(crc, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  CRCType remainder = UndoFinalize<CRCType, CRCWidth>(
+    crc, parameters.finalXOR,
+    parameters.reflectInput !=
+    parameters.reflectOutput);
 
-    remainder = CalculateRemainder(data, size, lookupTable, remainder);
+  remainder = CalculateRemainder(data, size, lookupTable, remainder);
 
-    // No need to mask the remainder here; the mask will be applied in the Finalize() function.
+  // No need to mask the remainder here; the mask will be applied in the Finalize() function.
 
-    return Finalize<CRCType, CRCWidth>(remainder, parameters.finalXOR, parameters.reflectInput != parameters.reflectOutput);
+  return Finalize<CRCType, CRCWidth>(
+    remainder, parameters.finalXOR,
+    parameters.reflectInput != parameters.reflectOutput);
 }
 
 /**
@@ -515,18 +549,17 @@ inline CRCType CRC::Calculate(const void * data, crcpp_size size, const Table<CR
     @tparam IntegerType Integer type of the value being reflected
     @return Reflected value
 */
-template <typename IntegerType>
+template<typename IntegerType>
 inline IntegerType CRC::Reflect(IntegerType value, crcpp_uint16 numBits)
 {
-    IntegerType reversedValue(0);
+  IntegerType reversedValue(0);
 
-    for (crcpp_uint16 i = 0; i < numBits; ++i)
-    {
-        reversedValue = static_cast<IntegerType>((reversedValue << 1) | (value & 1));
-        value = static_cast<IntegerType>(value >> 1);
-    }
+  for (crcpp_uint16 i = 0; i < numBits; ++i) {
+    reversedValue = static_cast<IntegerType>((reversedValue << 1) | (value & 1));
+    value = static_cast<IntegerType>(value >> 1);
+  }
 
-    return reversedValue;
+  return reversedValue;
 }
 
 /**
@@ -538,25 +571,26 @@ inline IntegerType CRC::Reflect(IntegerType value, crcpp_uint16 numBits)
     @tparam CRCWidth Number of bits in the CRC
     @return Final CRC
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline CRCType CRC::Finalize(CRCType remainder, CRCType finalXOR, bool reflectOutput)
 {
-    // For masking off the bits for the CRC (in the event that the number of bits in CRCType is larger than CRCWidth)
-    static crcpp_constexpr CRCType BIT_MASK = (CRCType(1) << (CRCWidth - CRCType(1))) |
-                                             ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1));
+  // For masking off the bits for the CRC
+  // (in the event that the number of bits in CRCType is larger than CRCWidth)
+  static crcpp_constexpr CRCType BIT_MASK = (CRCType(1) << (CRCWidth - CRCType(1))) |
+    ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1));
 
-    if (reflectOutput)
-    {
-        remainder = Reflect(remainder, CRCWidth);
-    }
+  if (reflectOutput) {
+    remainder = Reflect(remainder, CRCWidth);
+  }
 
-    return (remainder ^ finalXOR) & BIT_MASK;
+  return (remainder ^ finalXOR) & BIT_MASK;
 }
 
 /**
     @brief Undoes the process of computing the final reflection and XOR of a CRC remainder.
     @note This function allows for computation of multi-part CRCs
-    @note Calling UndoFinalize() followed by Finalize() (or vice versa) will always return the original remainder value:
+    @note Calling UndoFinalize() followed by Finalize() (or vice versa)
+          will always return the original remainder value:
 
         CRCType x = ...;
         CRCType y = Finalize(x, finalXOR, reflectOutput);
@@ -570,21 +604,21 @@ inline CRCType CRC::Finalize(CRCType remainder, CRCType finalXOR, bool reflectOu
     @tparam CRCWidth Number of bits in the CRC
     @return Un-finalized CRC remainder
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
+template<typename CRCType, crcpp_uint16 CRCWidth>
 inline CRCType CRC::UndoFinalize(CRCType crc, CRCType finalXOR, bool reflectOutput)
 {
-    // For masking off the bits for the CRC (in the event that the number of bits in CRCType is larger than CRCWidth)
-    static crcpp_constexpr CRCType BIT_MASK = (CRCType(1) << (CRCWidth - CRCType(1))) |
-                                             ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1));
+  // For masking off the bits for the CRC
+  // (in the event that the number of bits in CRCType is larger than CRCWidth)
+  static crcpp_constexpr CRCType BIT_MASK = (CRCType(1) << (CRCWidth - CRCType(1))) |
+    ((CRCType(1) << (CRCWidth - CRCType(1))) - CRCType(1));
 
-    crc = (crc & BIT_MASK) ^ finalXOR;
+  crc = (crc & BIT_MASK) ^ finalXOR;
 
-    if (reflectOutput)
-    {
-        crc = Reflect(crc, CRCWidth);
-    }
+  if (reflectOutput) {
+    crc = Reflect(crc, CRCWidth);
+  }
 
-    return crc;
+  return crc;
 }
 
 /**
@@ -597,111 +631,122 @@ inline CRCType CRC::UndoFinalize(CRCType crc, CRCType finalXOR, bool reflectOutp
     @tparam CRCWidth Number of bits in the CRC
     @return CRC remainder
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::CalculateRemainder(const void * data, crcpp_size size, const Parameters<CRCType, CRCWidth> & parameters, CRCType remainder)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::CalculateRemainder(
+  const void * data, crcpp_size size, const Parameters<CRCType,
+  CRCWidth> & parameters, CRCType remainder)
 {
 #ifdef CRCPP_USE_CPP11
-    // This static_assert is put here because this function will always be compiled in no matter what
-    // the template parameters are and whether or not a table lookup or bit-by-bit algorithm is used.
-    static_assert(::std::numeric_limits<CRCType>::digits >= CRCWidth, "CRCType is too small to contain a CRC of width CRCWidth.");
+  // This static_assert is put here because this function will always be compiled in no matter what
+  // the template parameters are and whether or not a table lookup or bit-by-bit algorithm is used.
+  static_assert(
+    ::std::numeric_limits<CRCType>::digits >= CRCWidth,
+    "CRCType is too small to contain a CRC of width CRCWidth.");
 #else
-    // Catching this compile-time error is very important. Sadly, the compiler error will be very cryptic, but it's
-    // better than nothing.
-    enum { static_assert_failed_CRCType_is_too_small_to_contain_a_CRC_of_width_CRCWidth = 1 / (::std::numeric_limits<CRCType>::digits >= CRCWidth ? 1 : 0) };
+  // Catching this compile-time error is very important. Sadly,
+  // the compiler error will be very cryptic, but it's better than nothing.
+  // cppcheck-suppress zerodiv
+  enum { static_assert_failed_CRCType_is_too_small_to_contain_a_CRC_of_width_CRCWidth = 1 /
+      (::std::numeric_limits<CRCType>::digits >= CRCWidth ? 1 : 0) };
 #endif
 
-    const unsigned char * current = reinterpret_cast<const unsigned char *>(data);
+  const unsigned char * current = reinterpret_cast<const unsigned char *>(data);
 
-    // Slightly different implementations based on the parameters. The current implementations try to eliminate as much
-    // computation from the inner loop (looping over each bit) as possible.
-    if (parameters.reflectInput)
-    {
-        CRCType polynomial = CRC::Reflect(parameters.polynomial, CRCWidth);
-        while (size--)
-        {
-            remainder = static_cast<CRCType>(remainder ^ *current++);
+  // Slightly different implementations based on the parameters.
+  // The current implementations try to eliminate as much
+  // computation from the inner loop (looping over each bit) as possible.
+  if (parameters.reflectInput) {
+    CRCType polynomial = CRC::Reflect(parameters.polynomial, CRCWidth);
+    while (size--) {
+      remainder = static_cast<CRCType>(remainder ^ *current++);
 
-            // An optimizing compiler might choose to unroll this loop.
-            for (crcpp_size i = 0; i < CHAR_BIT; ++i)
-            {
+      // An optimizing compiler might choose to unroll this loop.
+      for (crcpp_size i = 0; i < CHAR_BIT; ++i) {
 #ifdef CRCPP_BRANCHLESS
-                // Clever way to avoid a branch at the expense of a multiplication. This code is equivalent to the following:
-                // if (remainder & 1)
-                //     remainder = (remainder >> 1) ^ polynomial;
-                // else
-                //     remainder >>= 1;
-                remainder = static_cast<CRCType>((remainder >> 1) ^ ((remainder & 1) * polynomial));
+        // Clever way to avoid a branch at the expense of a multiplication.
+        // This code is equivalent to the following:
+        // if (remainder & 1)
+        //     remainder = (remainder >> 1) ^ polynomial;
+        // else
+        //     remainder >>= 1;
+        remainder = static_cast<CRCType>((remainder >> 1) ^ ((remainder & 1) * polynomial));
 #else
-                remainder = static_cast<CRCType>((remainder & 1) ? ((remainder >> 1) ^ polynomial) : (remainder >> 1));
+        remainder =
+          static_cast<CRCType>((remainder & 1) ? ((remainder >> 1) ^ polynomial) : (remainder >>
+          1));
 #endif
-            }
-        }
+      }
     }
-    else if (CRCWidth >= CHAR_BIT)
-    {
-        static crcpp_constexpr CRCType CRC_WIDTH_MINUS_ONE(CRCWidth - CRCType(1));
+  } else if (CRCWidth >= CHAR_BIT) {
+    static crcpp_constexpr CRCType CRC_WIDTH_MINUS_ONE(CRCWidth - CRCType(1));
 #ifndef CRCPP_BRANCHLESS
-        static crcpp_constexpr CRCType CRC_HIGHEST_BIT_MASK(CRCType(1) << CRC_WIDTH_MINUS_ONE);
+    static crcpp_constexpr CRCType CRC_HIGHEST_BIT_MASK(CRCType(1) << CRC_WIDTH_MINUS_ONE);
 #endif
-        // The conditional expression is used to avoid a -Wshift-count-overflow warning.
-        static crcpp_constexpr CRCType SHIFT((CRCWidth >= CHAR_BIT) ? static_cast<CRCType>(CRCWidth - CHAR_BIT) : 0);
+    // The conditional expression is used to avoid a -Wshift-count-overflow warning.
+    static crcpp_constexpr CRCType SHIFT(
+      (CRCWidth >= CHAR_BIT) ? static_cast<CRCType>(CRCWidth - CHAR_BIT) : 0);
 
-        while (size--)
-        {
-            remainder = static_cast<CRCType>(remainder ^ (static_cast<CRCType>(*current++) << SHIFT));
+    while (size--) {
+      remainder = static_cast<CRCType>(remainder ^ (static_cast<CRCType>(*current++) << SHIFT));
 
-            // An optimizing compiler might choose to unroll this loop.
-            for (crcpp_size i = 0; i < CHAR_BIT; ++i)
-            {
+      // An optimizing compiler might choose to unroll this loop.
+      for (crcpp_size i = 0; i < CHAR_BIT; ++i) {
 #ifdef CRCPP_BRANCHLESS
-                // Clever way to avoid a branch at the expense of a multiplication. This code is equivalent to the following:
-                // if (remainder & CRC_HIGHEST_BIT_MASK)
-                //     remainder = (remainder << 1) ^ parameters.polynomial;
-                // else
-                //     remainder <<= 1;
-                remainder = static_cast<CRCType>((remainder << 1) ^ (((remainder >> CRC_WIDTH_MINUS_ONE) & 1) * parameters.polynomial));
+        // Clever way to avoid a branch at the expense of a multiplication.
+        // This code is equivalent to the following:
+        // if (remainder & CRC_HIGHEST_BIT_MASK)
+        //     remainder = (remainder << 1) ^ parameters.polynomial;
+        // else
+        //     remainder <<= 1;
+        remainder =
+          static_cast<CRCType>((remainder <<
+          1) ^ (((remainder >> CRC_WIDTH_MINUS_ONE) & 1) * parameters.polynomial));
 #else
-                remainder = static_cast<CRCType>((remainder & CRC_HIGHEST_BIT_MASK) ? ((remainder << 1) ^ parameters.polynomial) : (remainder << 1));
+        remainder =
+          static_cast<CRCType>((remainder &
+          CRC_HIGHEST_BIT_MASK) ? ((remainder << 1) ^ parameters.polynomial) : (remainder << 1));
 #endif
-            }
-        }
+      }
     }
-    else
-    {
-        static crcpp_constexpr CRCType CHAR_BIT_MINUS_ONE(CHAR_BIT - 1);
+  } else {
+    static crcpp_constexpr CRCType CHAR_BIT_MINUS_ONE(CHAR_BIT - 1);
 #ifndef CRCPP_BRANCHLESS
-        static crcpp_constexpr CRCType CHAR_BIT_HIGHEST_BIT_MASK(CRCType(1) << CHAR_BIT_MINUS_ONE);
+    static crcpp_constexpr CRCType CHAR_BIT_HIGHEST_BIT_MASK(CRCType(1) << CHAR_BIT_MINUS_ONE);
 #endif
-        // The conditional expression is used to avoid a -Wshift-count-overflow warning.
-        static crcpp_constexpr CRCType SHIFT((CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
+    // The conditional expression is used to avoid a -Wshift-count-overflow warning.
+    static crcpp_constexpr CRCType SHIFT(
+      (CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
 
-        CRCType polynomial = static_cast<CRCType>(parameters.polynomial << SHIFT);
-        remainder = static_cast<CRCType>(remainder << SHIFT);
+    CRCType polynomial = static_cast<CRCType>(parameters.polynomial << SHIFT);
+    remainder = static_cast<CRCType>(remainder << SHIFT);
 
-        while (size--)
-        {
-            remainder = static_cast<CRCType>(remainder ^ *current++);
+    while (size--) {
+      remainder = static_cast<CRCType>(remainder ^ *current++);
 
-            // An optimizing compiler might choose to unroll this loop.
-            for (crcpp_size i = 0; i < CHAR_BIT; ++i)
-            {
+      // An optimizing compiler might choose to unroll this loop.
+      for (crcpp_size i = 0; i < CHAR_BIT; ++i) {
 #ifdef CRCPP_BRANCHLESS
-                // Clever way to avoid a branch at the expense of a multiplication. This code is equivalent to the following:
-                // if (remainder & CHAR_BIT_HIGHEST_BIT_MASK)
-                //     remainder = (remainder << 1) ^ polynomial;
-                // else
-                //     remainder <<= 1;
-                remainder = static_cast<CRCType>((remainder << 1) ^ (((remainder >> CHAR_BIT_MINUS_ONE) & 1) * polynomial));
+        // Clever way to avoid a branch at the expense of a multiplication.
+        // This code is equivalent to the following:
+        // if (remainder & CHAR_BIT_HIGHEST_BIT_MASK)
+        //     remainder = (remainder << 1) ^ polynomial;
+        // else
+        //     remainder <<= 1;
+        remainder =
+          static_cast<CRCType>((remainder <<
+          1) ^ (((remainder >> CHAR_BIT_MINUS_ONE) & 1) * polynomial));
 #else
-                remainder = static_cast<CRCType>((remainder & CHAR_BIT_HIGHEST_BIT_MASK) ? ((remainder << 1) ^ polynomial) : (remainder << 1));
+        remainder =
+          static_cast<CRCType>((remainder &
+          CHAR_BIT_HIGHEST_BIT_MASK) ? ((remainder << 1) ^ polynomial) : (remainder << 1));
 #endif
-            }
-        }
-
-        remainder = static_cast<CRCType>(remainder >> SHIFT);
+      }
     }
 
-    return remainder;
+    remainder = static_cast<CRCType>(remainder >> SHIFT);
+  }
+
+  return remainder;
 }
 
 /**
@@ -709,60 +754,61 @@ inline CRCType CRC::CalculateRemainder(const void * data, crcpp_size size, const
     @param[in] data Data over which the remainder will be computed
     @param[in] size Size of the data
     @param[in] lookupTable CRC lookup table
-    @param[in] remainder Running CRC remainder. Can be an initial value or the result of a previous CRC remainder calculation.
+    @param[in] remainder Running CRC remainder.
+               Can be an initial value or the result of a previous CRC remainder calculation.
     @tparam CRCType Integer type for storing the CRC result
     @tparam CRCWidth Number of bits in the CRC
     @return CRC remainder
 */
-template <typename CRCType, crcpp_uint16 CRCWidth>
-inline CRCType CRC::CalculateRemainder(const void * data, crcpp_size size, const Table<CRCType, CRCWidth> & lookupTable, CRCType remainder)
+template<typename CRCType, crcpp_uint16 CRCWidth>
+inline CRCType CRC::CalculateRemainder(
+  const void * data, crcpp_size size, const Table<CRCType,
+  CRCWidth> & lookupTable, CRCType remainder)
 {
-    const unsigned char * current = reinterpret_cast<const unsigned char *>(data);
+  const unsigned char * current = reinterpret_cast<const unsigned char *>(data);
 
-    if (lookupTable.GetParameters().reflectInput)
-    {
-        while (size--)
-        {
+  if (lookupTable.GetParameters().reflectInput) {
+    while (size--) {
 #if defined(WIN32) || defined(_WIN32) || defined(WINCE)
-    // Disable warning about data loss when doing (remainder >> CHAR_BIT) when
-    // remainder is one byte long. The algorithm is still correct in this case,
-    // though it's possible that one additional machine instruction will be executed.
+      // Disable warning about data loss when doing (remainder >> CHAR_BIT) when
+      // remainder is one byte long. The algorithm is still correct in this case,
+      // though it's possible that one additional machine instruction will be executed.
 #   pragma warning (push)
 #   pragma warning (disable : 4333)
 #endif
-            remainder = static_cast<CRCType>((remainder >> CHAR_BIT) ^ lookupTable[static_cast<unsigned char>(remainder ^ *current++)]);
+      remainder =
+        static_cast<CRCType>((remainder >>
+        CHAR_BIT) ^ lookupTable[static_cast<unsigned char>(remainder ^ *current++)]);
 #if defined(WIN32) || defined(_WIN32) || defined(WINCE)
 #   pragma warning (pop)
 #endif
-        }
     }
-    else if (CRCWidth >= CHAR_BIT)
-    {
-        // The conditional expression is used to avoid a -Wshift-count-overflow warning.
-        static crcpp_constexpr CRCType SHIFT((CRCWidth >= CHAR_BIT) ? static_cast<CRCType>(CRCWidth - CHAR_BIT) : 0);
+  } else if (CRCWidth >= CHAR_BIT) {
+    // The conditional expression is used to avoid a -Wshift-count-overflow warning.
+    static crcpp_constexpr CRCType SHIFT(
+      (CRCWidth >= CHAR_BIT) ? static_cast<CRCType>(CRCWidth - CHAR_BIT) : 0);
 
-        while (size--)
-        {
-            remainder = static_cast<CRCType>((remainder << CHAR_BIT) ^ lookupTable[static_cast<unsigned char>((remainder >> SHIFT) ^ *current++)]);
-        }
+    while (size--) {
+      remainder =
+        static_cast<CRCType>((remainder <<
+        CHAR_BIT) ^ lookupTable[static_cast<unsigned char>((remainder >> SHIFT) ^ *current++)]);
     }
-    else
-    {
-        // The conditional expression is used to avoid a -Wshift-count-overflow warning.
-        static crcpp_constexpr CRCType SHIFT((CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
+  } else {
+    // The conditional expression is used to avoid a -Wshift-count-overflow warning.
+    static crcpp_constexpr CRCType SHIFT(
+      (CHAR_BIT >= CRCWidth) ? static_cast<CRCType>(CHAR_BIT - CRCWidth) : 0);
 
-        remainder = static_cast<CRCType>(remainder << SHIFT);
+    remainder = static_cast<CRCType>(remainder << SHIFT);
 
-        while (size--)
-        {
-            // Note: no need to mask here since remainder is guaranteed to fit in a single byte.
-            remainder = lookupTable[static_cast<unsigned char>(remainder ^ *current++)];
-        }
-
-        remainder = static_cast<CRCType>(remainder >> SHIFT);
+    while (size--) {
+      // Note: no need to mask here since remainder is guaranteed to fit in a single byte.
+      remainder = lookupTable[static_cast<unsigned char>(remainder ^ *current++)];
     }
 
-    return remainder;
+    remainder = static_cast<CRCType>(remainder >> SHIFT);
+  }
+
+  return remainder;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -780,8 +826,8 @@ inline CRCType CRC::CalculateRemainder(const void * data, crcpp_size size, const
 */
 inline const CRC::Parameters<crcpp_uint8, 4> & CRC::CRC_4_ITU()
 {
-    static const Parameters<crcpp_uint8, 4> parameters = { 0x3, 0x0, 0x0, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 4> parameters = {0x3, 0x0, 0x0, true, true};
+  return parameters;
 }
 
 /**
@@ -798,8 +844,8 @@ inline const CRC::Parameters<crcpp_uint8, 4> & CRC::CRC_4_ITU()
 */
 inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_EPC()
 {
-    static const Parameters<crcpp_uint8, 5> parameters = { 0x09, 0x09, 0x00, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint8, 5> parameters = {0x09, 0x09, 0x00, false, false};
+  return parameters;
 }
 
 /**
@@ -816,8 +862,8 @@ inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_EPC()
 */
 inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_ITU()
 {
-    static const Parameters<crcpp_uint8, 5> parameters = { 0x15, 0x00, 0x00, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 5> parameters = {0x15, 0x00, 0x00, true, true};
+  return parameters;
 }
 
 /**
@@ -834,8 +880,8 @@ inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_ITU()
 */
 inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_USB()
 {
-    static const Parameters<crcpp_uint8, 5> parameters = { 0x05, 0x1F, 0x1F, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 5> parameters = {0x05, 0x1F, 0x1F, true, true};
+  return parameters;
 }
 
 /**
@@ -852,8 +898,8 @@ inline const CRC::Parameters<crcpp_uint8, 5> & CRC::CRC_5_USB()
 */
 inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_CDMA2000A()
 {
-    static const Parameters<crcpp_uint8, 6> parameters = { 0x27, 0x3F, 0x00, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint8, 6> parameters = {0x27, 0x3F, 0x00, false, false};
+  return parameters;
 }
 
 /**
@@ -870,8 +916,8 @@ inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_CDMA2000A()
 */
 inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_CDMA2000B()
 {
-    static const Parameters<crcpp_uint8, 6> parameters = { 0x07, 0x3F, 0x00, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint8, 6> parameters = {0x07, 0x3F, 0x00, false, false};
+  return parameters;
 }
 
 /**
@@ -888,8 +934,8 @@ inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_CDMA2000B()
 */
 inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_ITU()
 {
-    static const Parameters<crcpp_uint8, 6> parameters = { 0x03, 0x00, 0x00, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 6> parameters = {0x03, 0x00, 0x00, true, true};
+  return parameters;
 }
 
 /**
@@ -906,10 +952,10 @@ inline const CRC::Parameters<crcpp_uint8, 6> & CRC::CRC_6_ITU()
 */
 inline const CRC::Parameters<crcpp_uint8, 7> & CRC::CRC_7()
 {
-    static const Parameters<crcpp_uint8, 7> parameters = { 0x09, 0x00, 0x00, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint8, 7> parameters = {0x09, 0x00, 0x00, false, false};
+  return parameters;
 }
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 /**
     @brief Returns a set of parameters for CRC-8 SMBus.
@@ -925,8 +971,8 @@ inline const CRC::Parameters<crcpp_uint8, 7> & CRC::CRC_7()
 */
 inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8()
 {
-    static const Parameters<crcpp_uint8, 8> parameters = { 0x07, 0x00, 0x00, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint8, 8> parameters = {0x07, 0x00, 0x00, false, false};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -944,8 +990,8 @@ inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8()
 */
 inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_EBU()
 {
-    static const Parameters<crcpp_uint8, 8> parameters = { 0x1D, 0xFF, 0x00, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 8> parameters = {0x1D, 0xFF, 0x00, true, true};
+  return parameters;
 }
 
 /**
@@ -962,8 +1008,8 @@ inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_EBU()
 */
 inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_MAXIM()
 {
-    static const Parameters<crcpp_uint8, 8> parameters = { 0x31, 0x00, 0x00, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 8> parameters = {0x31, 0x00, 0x00, true, true};
+  return parameters;
 }
 
 /**
@@ -980,8 +1026,8 @@ inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_MAXIM()
 */
 inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_WCDMA()
 {
-    static const Parameters<crcpp_uint8, 8> parameters = { 0x9B, 0x00, 0x00, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint8, 8> parameters = {0x9B, 0x00, 0x00, true, true};
+  return parameters;
 }
 
 /**
@@ -998,8 +1044,8 @@ inline const CRC::Parameters<crcpp_uint8, 8> & CRC::CRC_8_WCDMA()
 */
 inline const CRC::Parameters<crcpp_uint16, 10> & CRC::CRC_10()
 {
-    static const Parameters<crcpp_uint16, 10> parameters = { 0x233, 0x000, 0x000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 10> parameters = {0x233, 0x000, 0x000, false, false};
+  return parameters;
 }
 
 /**
@@ -1016,8 +1062,8 @@ inline const CRC::Parameters<crcpp_uint16, 10> & CRC::CRC_10()
 */
 inline const CRC::Parameters<crcpp_uint16, 10> & CRC::CRC_10_CDMA2000()
 {
-    static const Parameters<crcpp_uint16, 10> parameters = { 0x3D9, 0x3FF, 0x000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 10> parameters = {0x3D9, 0x3FF, 0x000, false, false};
+  return parameters;
 }
 
 /**
@@ -1034,8 +1080,8 @@ inline const CRC::Parameters<crcpp_uint16, 10> & CRC::CRC_10_CDMA2000()
 */
 inline const CRC::Parameters<crcpp_uint16, 11> & CRC::CRC_11()
 {
-    static const Parameters<crcpp_uint16, 11> parameters = { 0x385, 0x01A, 0x000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 11> parameters = {0x385, 0x01A, 0x000, false, false};
+  return parameters;
 }
 
 /**
@@ -1052,8 +1098,8 @@ inline const CRC::Parameters<crcpp_uint16, 11> & CRC::CRC_11()
 */
 inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_CDMA2000()
 {
-    static const Parameters<crcpp_uint16, 12> parameters = { 0xF13, 0xFFF, 0x000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 12> parameters = {0xF13, 0xFFF, 0x000, false, false};
+  return parameters;
 }
 
 /**
@@ -1070,8 +1116,8 @@ inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_CDMA2000()
 */
 inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_DECT()
 {
-    static const Parameters<crcpp_uint16, 12> parameters = { 0x80F, 0x000, 0x000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 12> parameters = {0x80F, 0x000, 0x000, false, false};
+  return parameters;
 }
 
 /**
@@ -1088,8 +1134,8 @@ inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_DECT()
 */
 inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_UMTS()
 {
-    static const Parameters<crcpp_uint16, 12> parameters = { 0x80F, 0x000, 0x000, false, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 12> parameters = {0x80F, 0x000, 0x000, false, true};
+  return parameters;
 }
 
 /**
@@ -1106,8 +1152,8 @@ inline const CRC::Parameters<crcpp_uint16, 12> & CRC::CRC_12_UMTS()
 */
 inline const CRC::Parameters<crcpp_uint16, 13> & CRC::CRC_13_BBC()
 {
-    static const Parameters<crcpp_uint16, 13> parameters = { 0x1CF5, 0x0000, 0x0000, false, false };
-   return parameters;
+  static const Parameters<crcpp_uint16, 13> parameters = {0x1CF5, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1124,8 +1170,8 @@ inline const CRC::Parameters<crcpp_uint16, 13> & CRC::CRC_13_BBC()
 */
 inline const CRC::Parameters<crcpp_uint16, 15> & CRC::CRC_15()
 {
-    static const Parameters<crcpp_uint16, 15> parameters = { 0x4599, 0x0000, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 15> parameters = {0x4599, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1142,10 +1188,10 @@ inline const CRC::Parameters<crcpp_uint16, 15> & CRC::CRC_15()
 */
 inline const CRC::Parameters<crcpp_uint16, 15> & CRC::CRC_15_MPT1327()
 {
-    static const Parameters<crcpp_uint16, 15> parameters = { 0x6815, 0x0000, 0x0001, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 15> parameters = {0x6815, 0x0000, 0x0001, false, false};
+  return parameters;
 }
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 /**
     @brief Returns a set of parameters for CRC-16 ARC (aka CRC-16 IBM, CRC-16 LHA).
@@ -1161,8 +1207,8 @@ inline const CRC::Parameters<crcpp_uint16, 15> & CRC::CRC_15_MPT1327()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_ARC()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0x0000, 0x0000, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0x0000, 0x0000, true, true};
+  return parameters;
 }
 
 /**
@@ -1179,8 +1225,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_ARC()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_BUYPASS()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0x0000, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1197,8 +1243,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_BUYPASS()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CCITTFALSE()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x1021, 0xFFFF, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x1021, 0xFFFF, 0x0000, false, false};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -1216,8 +1262,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CCITTFALSE()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CDMA2000()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0xC867, 0xFFFF, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0xC867, 0xFFFF, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1234,8 +1280,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CDMA2000()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CMS()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0xFFFF, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0xFFFF, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1252,8 +1298,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_CMS()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DECTR()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x0589, 0x0000, 0x0001, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x0589, 0x0000, 0x0001, false, false};
+  return parameters;
 }
 
 /**
@@ -1270,8 +1316,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DECTR()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DECTX()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x0589, 0x0000, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x0589, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1288,10 +1334,10 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DECTX()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DNP()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x3D65, 0x0000, 0xFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x3D65, 0x0000, 0xFFFF, true, true};
+  return parameters;
 }
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 /**
     @brief Returns a set of parameters for CRC-16 GENIBUS (aka CRC-16 EPC, CRC-16 I-CODE, CRC-16 DARC).
@@ -1307,8 +1353,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_DNP()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_GENIBUS()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x1021, 0xFFFF, 0xFFFF, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x1021, 0xFFFF, 0xFFFF, false, false};
+  return parameters;
 }
 
 /**
@@ -1325,8 +1371,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_GENIBUS()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_KERMIT()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x1021, 0x0000, 0x0000, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x1021, 0x0000, 0x0000, true, true};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -1344,8 +1390,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_KERMIT()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_MAXIM()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0x0000, 0xFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0x0000, 0xFFFF, true, true};
+  return parameters;
 }
 
 /**
@@ -1362,8 +1408,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_MAXIM()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_MODBUS()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0xFFFF, 0x0000, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0xFFFF, 0x0000, true, true};
+  return parameters;
 }
 
 /**
@@ -1380,8 +1426,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_MODBUS()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_T10DIF()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8BB7, 0x0000, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8BB7, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 /**
@@ -1398,11 +1444,11 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_T10DIF()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_USB()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x8005, 0xFFFF, 0xFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x8005, 0xFFFF, 0xFFFF, true, true};
+  return parameters;
 }
 
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 /**
     @brief Returns a set of parameters for CRC-16 X-25 (aka CRC-16 IBM-SDLC, CRC-16 ISO-HDLC, CRC-16 B).
@@ -1418,8 +1464,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_USB()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_X25()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x1021, 0xFFFF, 0xFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x1021, 0xFFFF, 0xFFFF, true, true};
+  return parameters;
 }
 
 /**
@@ -1436,8 +1482,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_X25()
 */
 inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_XMODEM()
 {
-    static const Parameters<crcpp_uint16, 16> parameters = { 0x1021, 0x0000, 0x0000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint16, 16> parameters = {0x1021, 0x0000, 0x0000, false, false};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -1455,8 +1501,8 @@ inline const CRC::Parameters<crcpp_uint16, 16> & CRC::CRC_16_XMODEM()
 */
 inline const CRC::Parameters<crcpp_uint32, 17> & CRC::CRC_17_CAN()
 {
-    static const Parameters<crcpp_uint32, 17> parameters = { 0x1685B, 0x00000, 0x00000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32, 17> parameters = {0x1685B, 0x00000, 0x00000, false, false};
+  return parameters;
 }
 
 /**
@@ -1473,8 +1519,9 @@ inline const CRC::Parameters<crcpp_uint32, 17> & CRC::CRC_17_CAN()
 */
 inline const CRC::Parameters<crcpp_uint32, 21> & CRC::CRC_21_CAN()
 {
-    static const Parameters<crcpp_uint32, 21> parameters = { 0x102899, 0x000000, 0x000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    21> parameters = {0x102899, 0x000000, 0x000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1491,8 +1538,9 @@ inline const CRC::Parameters<crcpp_uint32, 21> & CRC::CRC_21_CAN()
 */
 inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24()
 {
-    static const Parameters<crcpp_uint32, 24> parameters = { 0x864CFB, 0xB704CE, 0x000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    24> parameters = {0x864CFB, 0xB704CE, 0x000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1509,8 +1557,9 @@ inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24()
 */
 inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24_FLEXRAYA()
 {
-    static const Parameters<crcpp_uint32, 24> parameters = { 0x5D6DCB, 0xFEDCBA, 0x000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    24> parameters = {0x5D6DCB, 0xFEDCBA, 0x000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1527,8 +1576,9 @@ inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24_FLEXRAYA()
 */
 inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24_FLEXRAYB()
 {
-    static const Parameters<crcpp_uint32, 24> parameters = { 0x5D6DCB, 0xABCDEF, 0x000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    24> parameters = {0x5D6DCB, 0xABCDEF, 0x000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1545,10 +1595,11 @@ inline const CRC::Parameters<crcpp_uint32, 24> & CRC::CRC_24_FLEXRAYB()
 */
 inline const CRC::Parameters<crcpp_uint32, 30> & CRC::CRC_30()
 {
-    static const Parameters<crcpp_uint32, 30> parameters = { 0x2030B9C7, 0x3FFFFFFF, 0x00000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    30> parameters = {0x2030B9C7, 0x3FFFFFFF, 0x00000000, false, false};
+  return parameters;
 }
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 /**
     @brief Returns a set of parameters for CRC-32 (aka CRC-32 ADCCP, CRC-32 PKZip).
@@ -1564,8 +1615,9 @@ inline const CRC::Parameters<crcpp_uint32, 30> & CRC::CRC_30()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x04C11DB7, 0xFFFFFFFF, 0xFFFFFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x04C11DB7, 0xFFFFFFFF, 0xFFFFFFFF, true, true};
+  return parameters;
 }
 
 /**
@@ -1582,8 +1634,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_BZIP2()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x04C11DB7, 0xFFFFFFFF, 0xFFFFFFFF, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x04C11DB7, 0xFFFFFFFF, 0xFFFFFFFF, false, false};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -1601,8 +1654,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_BZIP2()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_C()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x1EDC6F41, 0xFFFFFFFF, 0xFFFFFFFF, true, true };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x1EDC6F41, 0xFFFFFFFF, 0xFFFFFFFF, true, true};
+  return parameters;
 }
 #endif
 
@@ -1620,8 +1674,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_C()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_MPEG2()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x04C11DB7, 0xFFFFFFFF, 0x00000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x04C11DB7, 0xFFFFFFFF, 0x00000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1638,8 +1693,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_MPEG2()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_POSIX()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x04C11DB7, 0x00000000, 0xFFFFFFFF, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x04C11DB7, 0x00000000, 0xFFFFFFFF, false, false};
+  return parameters;
 }
 
 #ifdef CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
@@ -1657,8 +1713,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_POSIX()
 */
 inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_Q()
 {
-    static const Parameters<crcpp_uint32, 32> parameters = { 0x814141AB, 0x00000000, 0x00000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint32,
+    32> parameters = {0x814141AB, 0x00000000, 0x00000000, false, false};
+  return parameters;
 }
 
 /**
@@ -1675,8 +1732,9 @@ inline const CRC::Parameters<crcpp_uint32, 32> & CRC::CRC_32_Q()
 */
 inline const CRC::Parameters<crcpp_uint64, 40> & CRC::CRC_40_GSM()
 {
-    static const Parameters<crcpp_uint64, 40> parameters = { 0x0004820009, 0x0000000000, 0xFFFFFFFFFF, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint64,
+    40> parameters = {0x0004820009, 0x0000000000, 0xFFFFFFFFFF, false, false};
+  return parameters;
 }
 
 /**
@@ -1693,13 +1751,14 @@ inline const CRC::Parameters<crcpp_uint64, 40> & CRC::CRC_40_GSM()
 */
 inline const CRC::Parameters<crcpp_uint64, 64> & CRC::CRC_64()
 {
-    static const Parameters<crcpp_uint64, 64> parameters = { 0x42F0E1EBA9EA3693, 0x0000000000000000, 0x0000000000000000, false, false };
-    return parameters;
+  static const Parameters<crcpp_uint64,
+    64> parameters = {0x42F0E1EBA9EA3693, 0x0000000000000000, 0x0000000000000000, false, false};
+  return parameters;
 }
-#endif // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
+#endif  // CRCPP_INCLUDE_ESOTERIC_CRC_DEFINITIONS
 
 #ifdef CRCPP_USE_NAMESPACE
-}
+}  // namespace CRCPP
 #endif
 
-#endif // CRCPP_CRC_H_
+#endif  // VESC_DRIVER__CRC_HPP_
