@@ -1,32 +1,37 @@
 // Copyright 2020 F1TENTH Foundation
 //
-// Redistribution and use in source and binary forms, with or without modification, are permitted
-// provided that the following conditions are met:
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
 //
-// 1. Redistributions of source code must retain the above copyright notice, this list of conditions
-//    and the following disclaimer.
+//   * Redistributions of source code must retain the above copyright
+//     notice, this list of conditions and the following disclaimer.
 //
-// 2. Redistributions in binary form must reproduce the above copyright notice, this list
-//    of conditions and the following disclaimer in the documentation and/or other materials
-//    provided with the distribution.
+//   * Redistributions in binary form must reproduce the above copyright
+//     notice, this list of conditions and the following disclaimer in the
+//     documentation and/or other materials provided with the distribution.
 //
-// 3. Neither the name of the copyright holder nor the names of its contributors may be used
-//    to endorse or promote products derived from this software without specific prior
-//    written permission.
+//   * Neither the name of the {copyright_holder} nor the names of its
+//     contributors may be used to endorse or promote products derived from
+//     this software without specific prior written permission.
 //
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
-// IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
-// FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
-// WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
-// WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
 
 // -*- mode:c++; fill-column: 100; -*-
 
-#ifndef VESC_DRIVER_VESC_PACKET_FACTORY_H_
-#define VESC_DRIVER_VESC_PACKET_FACTORY_H_
+#ifndef VESC_DRIVER__VESC_PACKET_FACTORY_HPP_
+#define VESC_DRIVER__VESC_PACKET_FACTORY_HPP_
+
+#include "vesc_driver/vesc_packet.hpp"
 
 #include <cstdint>
 #include <functional>
@@ -34,8 +39,6 @@
 #include <memory>
 #include <string>
 #include <vector>
-
-#include "vesc_driver/vesc_packet.h"
 
 namespace vesc_driver
 {
@@ -47,7 +50,7 @@ class VescPacketFactory
 {
 public:
   /** Return the global factory object */
-  static VescPacketFactory* getFactory();
+  static VescPacketFactory * getFactory();
 
   /**
    * Create a VescPacket from a buffer (factory function). Packet must start (start of frame
@@ -69,9 +72,10 @@ public:
    *
    * @return Pointer to a valid VescPacket if successful. Otherwise, an empty pointer.
    */
-  static VescPacketPtr createPacket(const Buffer::const_iterator& begin,
-                                    const Buffer::const_iterator& end,
-                                    int* num_bytes_needed, std::string* what);
+  static VescPacketPtr createPacket(
+    const Buffer::const_iterator & begin,
+    const Buffer::const_iterator & end,
+    int * num_bytes_needed, std::string * what);
 
   typedef std::function<VescPacketPtr(std::shared_ptr<VescFrame>)> CreateFn;
 
@@ -86,8 +90,8 @@ public:
 
 private:
   VescPacketFactory();
-  typedef std::map<int, CreateFn > FactoryMap;
-  static FactoryMap* getMap();
+  typedef std::map<int, CreateFn> FactoryMap;
+  static FactoryMap * getMap();
 };
 
 template<typename PACKETTYPE>
@@ -107,8 +111,8 @@ public:
 
 /** Use this macro to register packets */
 #define REGISTER_PACKET_TYPE(id, klass) \
-static PacketFactoryTemplate<klass> global_##klass##Factory((id));
+  static PacketFactoryTemplate<klass> global_ ## klass ## Factory((id));
 
 }  // namespace vesc_driver
 
-#endif  // VESC_DRIVER_VESC_PACKET_FACTORY_H_
+#endif  // VESC_DRIVER__VESC_PACKET_FACTORY_HPP_
